refactor: money 객체를 도입하며 베팅금액 수정하는 기능 수정
