+++ src/main/java/blackjack/bettingMachine/Money.java
@@ -1,0 +1,14 @@
+package blackjack.bettingMachine;
+
+public class Money {
+
+    private long money;
+
+    public Money(final long money) {
+        this.money = money;
+    }
+
+    public void increase(final long amount) {
+        this.money += amount;
+    }
+}
